Add basic implementation of `length` assertions.
This closes #5. Tests now pass. More tests should be added, but I think
this is a reasonable implementation. There's nothing to do here, really.
Probably deciding which types to compile for and doing type conversions
is the real trouble with all functions in this library.

This also bumps version to 0.0.7.